Day 12 Part B Fix

+++ 2022/Day 12/b.cpp
@@ -34,7 +34,7 @@ vector<coord> getChildren(const coord& c) {
   return res;
 }
 
-size_t bfs() {
+int bfs() {
   map<coord, coord> childToParent;
   childToParent.insert({endCoord, endCoord});
